try to handle different chunks of data
handle things like send a chunk delay some time, send another and so on.
do not send a big chunk collapse multiple ones.

For instance if latency is set with 30ms and we send packets every 8ms
we should try to reply this instead of sending first, wait 8ms and send
one bigger packet.