initial client/server implementation
Allows to restrict latency/bandwidth between 2 machines.
This require to setup both machines pointing to each other.
Works but it's hard to change parameters.
To do that both end encapsulate packets in a UDP packet (the UDP data
it's just the packet received from tun interface). Quite easy but works.
Would be easier if server will be launched just with a port specification
while client with all parameters so to make easier to change.