Use 2 tun devices for local case
This is required to better handle 2 network queue (back and forth)
improving network emulation. In real case there are 2 queue, one on
local machine and another in the remote one.
This will also allow to reduce the queue.
Another problem is that using a single queue there was a single flow
and the combined bandwidth (forth and back) was limited using a single
flow. Now two flows are correctly used for the local case.
This require some more complicate NAT fixing the checksums.

Signed-off-by: Frediano Ziglio <fziglio@redhat.com>