Use ppoll and different threading
Instead of using a thread for reading from tun(s)/remote use a
thread for every direction.
This remove a lot of synchronisation from the code.
Some test I did prove that this implementation although using
a single thread for a direction is faster allowing to use higher
speed.
Also this make easier to reduce the internal queue at a minimum
emulating better a real card and allowing to do some tests with QoS
settings.

Signed-off-by: Frediano Ziglio <fziglio@redhat.com>